Major changes
Comment section at the top of the file explains how to structure the the csv file. 3/4 of the code has been moved into the get_link function that is now called from the line that launches the web browser. While loop also continues to display the formatted output to the screen unless the user chooses not to.